Refactor constants and update unit tests for motion photo version. Add Javadoc for constants classes.
