+++ MotionPhotoReader/app/src/main/java/com/google/android/libraries/motionphotoreader/Constants.java
@@ -5,41 +5,46 @@ package com.google.android.libraries.motionphotoreader;
  */
 class Constants {
 
-    /* *********************************************************************************************
-     * BufferProcessor.java
-     * ********************************************************************************************/
-
+    /**
+     * Time to wait for a blocking queue to offer buffer information when polled, in microseconds.
+     */
     static final long TIMEOUT_US = 1000L;
+
+    /**
+     * Conversion factor from microseconds to nanoseconds.
+     */
     static final long US_TO_NS = 1000L;
+
+    /**
+     * Use this time difference (approx. 30 FPS) between consecutive frames if a reasonable frame
+     * delta cannot be computed from the frame presentation timestamps.
+     */
     static final long FALLBACK_FRAME_DELTA_NS = 1_000_000_000L / 30;
+
+    /**
+     * The number of strips that each frame is divided into for stabilization. A separate
+     * stabilization homography is applied to each strip. The default number of strips is 12, but
+     * a single strip suffices in terms of visual results.
+     */
     static final int NUM_OF_STRIPS = 12;
 
-    /* *********************************************************************************************
-     * HomographyMatrix.java
-     * ********************************************************************************************/
-
+    /**
+     * Float array for the identity matrix.
+     */
     static final float[] IDENTITY = {
             1.0f, 0.0f, 0.0f,
             0.0f, 1.0f, 0.0f,
             0.0f, 0.0f, 1.0f
     };
 
+    /**
+     * Error threshold for two floats in a homography matrix to be considered equal.
+     */
     static final float EPS = 1E-3f;
-
-    /* *********************************************************************************************
-     * MotionPhotoInfo.java
-     * ********************************************************************************************/
 
     static final int MOTION_PHOTO_V1 = 1;
     static final int MOTION_PHOTO_V2 = 2;
     static final String CAMERA_XMP_NAMESPACE = "http://ns.google.com/photos/1.0/camera/";
-    static final String V2_XMP_PROP_PREFIX = "Container:Directory[";
-    static final String V2_XMP_PROP_LENGTH_SUFFIX = "]/Container:Item/Item:Length";
-    static final String V2_XMP_PROP_PADDING_SUFFIX = "]/Container:Item/Item:Length";
-
-    /* *********************************************************************************************
-     * MotionPhotoReader.java
-     * ********************************************************************************************/
 
     /**
      * String representing the MIME type for the track that contains information about the video
@@ -55,61 +60,21 @@ class Constants {
      */
     static final String VIDEO_MIME_PREFIX = "video/";
 
-    /* *********************************************************************************************
-     * TextureRender.java
-     * ********************************************************************************************/
-
+    /**
+     * Offsets and constants for OpenGL rendering.
+     */
     static final int FLOAT_SIZE_BYTES = 4;
     static final int TRIANGLE_VERTICES_DATA_POS_OFFSET = 0;
     static final int TRIANGLE_VERTICES_DATA_STRIDE_BYTES = 3 * FLOAT_SIZE_BYTES;
 
-    /* *********************************************************************************************
-     * XmpParser.java
-     * ********************************************************************************************/
-
-    static final byte[] OPEN_ARR = "<x:xmpmeta".getBytes();          /* Start of XMP metadata tag */
-    static final byte[] CLOSE_ARR = "</x:xmpmeta>".getBytes();       /* End of XMP metadata tag   */
-
-    /* *********************************************************************************************
-     * Tests
-     * ********************************************************************************************/
+    /**
+     * Strings demarcating start and end of XMP metadata tag.
+     */
+    static final byte[] OPEN_ARR = "<x:xmpmeta".getBytes();
+    static final byte[] CLOSE_ARR = "</x:xmpmeta>".getBytes();
 
     /**
-     * MotionPhotoReaderTest.java
+     * Asset directory containing all motion photo files.
      */
-    static final int NUM_FRAMES = 43;
-    static final long SEEK_AMOUNT_US = 10_000L;
-
-    /**
-     * MotionPhotoInfoTest.java
-     */
-    static final int KEY_WIDTH = 4032;
-    static final int KEY_HEIGHT = 3024;
-    static final String KEY_MIME = "video/avc";
-
-    // Define a mock media format for motion photo v1 format
-    static final String FILENAME_V1 = "MVIMG_20200621_200240.jpg";
-    static final long KEY_DURATION_V1 = 1499400;
-    static final int KEY_ROTATION_V1 = 90;
-    static final int VIDEO_OFFSET_V1 = 2592317;
-
-    // Define a mock media format for motion photo v2 format
-    static final String FILENAME_V2 = "PXL_20200710_061629144.MP.jpg";
-    static final long KEY_DURATION_V2 = 763422;
-    static final int KEY_ROTATION_V2 = 0;
-    static final int VIDEO_OFFSET_V2 = 1317283;
-
-    /**
-     * HomographyMatrix.java
-     */
-    static final double THETA_DEGREES_A = 45.0;
-    static final double THETA_DEGREES_B = 60.0;
-    static final int VIDEO_WIDTH_PIXELS = 4032;
-    static final int VIDEO_HEIGHT_PIXELS = 3024;
-
-    /* *********************************************************************************************
-     * Directory names.
-     * ********************************************************************************************/
-
     static final String MOTION_PHOTOS_DIR = "motionphotos/";
 }
